refactor(AuthController): Login 메서드에서 JWT token을 body에 담아 보내도록 변경
